test: 合并 bangumi_api/bangumi_data 家族
- 删 test_bangumi_api_full(5)/comprehensive(8) 全重复冒烟
- internals 6 独有断言并入 test_bangumi_api.py 后删除文件
- data_comprehensive 4 强断言 + data_internals 6 条并入 test_bangumi_data.py 后删除
- extended 删 11 条 hasattr 冒烟与重复的 GetMovieMainEpisodeId，独有链遍历用例保留
