--- tests/utils/test_bangumi_data_internals.py
@@ -1,90 +1,0 @@
-"""bangumi_data 模块级重试与 BangumiData 缓存/下载边界（mock，不依赖外网）。"""
-
-from unittest.mock import MagicMock, patch
-
-import httpx
-import pytest
-
-from app.utils import bangumi_data
-
-
-@patch("app.utils.bangumi_data.time.sleep")
-@patch("app.utils.bangumi_data.httpx.Client")
-def test_module_request_with_retry_raises_after_exhausted(mock_client_cls, _sleep):
-    mock_client = mock_client_cls.return_value
-    mock_client.request.side_effect = httpx.ConnectError("down")
-    with pytest.raises(httpx.ConnectError):
-        bangumi_data._request_with_retry(
-            "https://example.test/data.json", max_retries=1, ssl_verify=True
-        )
-    assert mock_client.request.call_count == 2
-
-
-@patch("app.utils.bangumi_data.time.sleep")
-@patch("app.utils.bangumi_data.httpx.Client")
-def test_module_request_with_retry_connection_error_then_success(
-    mock_client_cls, _sleep
-):
-    ok = MagicMock()
-    ok.status_code = 200
-    ok.raise_for_status = MagicMock()
-    ok.elapsed.total_seconds.return_value = 0.01
-    ok.headers = {}
-    ok.text = ""
-    mock_client = mock_client_cls.return_value
-    mock_client.request.side_effect = [
-        httpx.ConnectError("down"),
-        ok,
-    ]
-    out = bangumi_data._request_with_retry(
-        "https://example.test/retry.json", max_retries=2, ssl_verify=True
-    )
-    assert out.status_code == 200
-    assert mock_client.request.call_count == 2
-
-
-@patch("app.utils.bangumi_data.time.sleep")
-@patch("app.utils.bangumi_data.httpx.Client")
-def test_module_request_with_retry_ssl_verify_false_sets_warnings(
-    mock_client_cls, _sleep
-):
-    ok = MagicMock()
-    ok.status_code = 200
-    ok.raise_for_status = MagicMock()
-    ok.elapsed.total_seconds.return_value = 0.01
-    ok.headers = {}
-    ok.text = ""
-    mock_client = mock_client_cls.return_value
-    mock_client.request.return_value = ok
-    bangumi_data._request_with_retry(
-        "https://example.test/x", max_retries=0, ssl_verify=False
-    )
-    mock_client_cls.assert_called_once()
-
-
-@patch.object(bangumi_data.BangumiData, "_build_tmdb_mapping", lambda self: None)
-@patch.object(bangumi_data.BangumiData, "_preload_data_to_memory", lambda self: None)
-@patch.object(
-    bangumi_data.BangumiData,
-    "_check_and_download_cache_on_startup",
-    lambda self: None,
-)
-class TestBangumiDataCacheHelpers:
-    def test_is_cache_valid_false_when_missing(self):
-        data = bangumi_data.BangumiData()
-        with patch("os.path.exists", return_value=False):
-            assert data._is_cache_valid() is False
-
-    def test_is_cache_valid_false_on_mtime_error(self):
-        data = bangumi_data.BangumiData()
-        with patch("os.path.exists", return_value=True):
-            with patch("os.path.getmtime", side_effect=OSError("no mtime")):
-                assert data._is_cache_valid() is False
-
-    def test_download_data_returns_false_on_request_failure(self):
-        data = bangumi_data.BangumiData()
-        with patch(
-            "app.utils.bangumi_data._request_with_retry",
-            side_effect=httpx.HTTPError("fail"),
-        ):
-            assert data._download_data() is False
